Upgrade authentication session management
